Generalize ItemPickerMenu impl
